# REGR: Performance decrease in factorize