--- a/pandas/core/algorithms.py
+++ b/pandas/core/algorithms.py
@@ -810,11 +810,7 @@ def factorize(
             if null_mask.any():
                 na_value = na_value_for_dtype(values.dtype, compat=False)
                 # Don't modify (potentially user-provided) array
-                values = np.where(
-                    null_mask,
-                    na_value,
-                    values,
-                )
+                values = np.where(null_mask, na_value, values)
 
         codes, uniques = factorize_array(
             values,